backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt 
djoser
django-cors-headers
python-decouple
psycopg2
